small refactor for serve file
