--- judge/utils/user_upload_access.py
@@ -1,27 +1,0 @@
-import mimetypes
-from urllib.parse import quote
-
-from django.http import Http404, HttpResponse
-from django.utils.translation import gettext_lazy as _
-
-from judge.utils.views import add_file_response, generic_message
-
-
-def authorize_file_access(request, file_obj):
-    user = request.user
-    if user.is_authenticated and (user.is_superuser or file_obj.is_owned_by(user)):
-        return file_obj
-    raise Http404('File not found or access denied.')
-
-
-def serve_user_upload(request, user_upload):
-    """Build an inline HTTP response for a UserUpload, using X-Accel-Redirect when available."""
-    try:
-        response = HttpResponse()
-        response['Content-Type'] = mimetypes.guess_type(user_upload.filename)[0] or 'application/octet-stream'
-        encoded = quote(user_upload.filename, safe='')
-        response['Content-Disposition'] = f"inline; filename*=UTF-8''{encoded}"
-        add_file_response(request, response, user_upload.get_internal_url_path(), user_upload.get_file_path())
-        return response
-    except (OSError, IOError):
-        return generic_message(request, 'File Error', _('File not found.'), status=404)
